+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 8.1.4 (git master)
+* Import milestone constraints from Asta schedules (Contributed by Dave McKay)
+* Correctly determine the constraint type for tasks with ALAP placement with or without predecessors when reading from from Asta schedules (Contributed by Dave McKay)
 
 ## 8.1.3 (25/06/2020)
 * Improve reliability when reading custom field values from certain MPP12 files.